Module peepmatic_runtime::linear

A linear IR for optimizations.

This IR is designed such that it should be easy to combine multiple linear optimizations into a single automata.

See also src/linearize.rs for the AST to linear IR translation pass.

Structs

Else

Match any value.

LhsId

A canonicalized identifier for a left-hand side value that was bound in a pattern.

Match

A partial match of an optimization’s LHS.

Optimization

A linearized optimization.

Optimizations

A set of linear optimizations.

RhsId

A canonicalized identifier for a right-hand side value.

Enums

Action

An action to perform when transitioning between states in the automata.

MatchOp

A matching operation to be performed on some Cranelift instruction as part of determining whether an optimization is applicable.

Functions

bool_to_match_result

Convert a boolean to a MatchResult.

Type Definitions

MatchResult

The result of evaluating a MatchOp.