Trait peepmatic_runtime::operator::TypingContext

pub trait TypingContext<'a> {
    type TypeVariable;
    fn cc(&mut self, span: Span) -> Self::TypeVariable;

    fn bNN(&mut self, span: Span) -> Self::TypeVariable;

    fn iNN(&mut self, span: Span) -> Self::TypeVariable;

    fn iMM(&mut self, span: Span) -> Self::TypeVariable;

    fn cpu_flags(&mut self, span: Span) -> Self::TypeVariable;

    fn b1(&mut self, span: Span) -> Self::TypeVariable;

    fn void(&mut self, span: Span) -> Self::TypeVariable;

    fn bool_or_int(&mut self, span: Span) -> Self::TypeVariable;

    fn any_t(&mut self, span: Span) -> Self::TypeVariable;
}

A trait to represent a typing context.

This is used by the macro-generated operator methods that create the type variables for their immediates, parameters, and results. This trait is implemented by the concrete typing context in peepmatic/src/verify.rs.

Associated Types

type TypeVariable

A type variable.

Loading content...
Required methods

fn cc(&mut self, span: Span) -> Self::TypeVariable

Create a condition code type.

fn bNN(&mut self, span: Span) -> Self::TypeVariable

Create a boolean type with a polymorphic bit width.

Each use of bNN by the same operator refers to the same type variable.

fn iNN(&mut self, span: Span) -> Self::TypeVariable

Create an integer type with a polymorphic bit width.

Each use of iNN by the same operator refers to the same type variable.

fn iMM(&mut self, span: Span) -> Self::TypeVariable

Create an integer type with a polymorphic bit width.

Each use of iMM by the same operator refers to the same type variable.

fn cpu_flags(&mut self, span: Span) -> Self::TypeVariable

Create the CPU flags type variable.

fn b1(&mut self, span: Span) -> Self::TypeVariable

Create a boolean type of size one bit.

fn void(&mut self, span: Span) -> Self::TypeVariable

Create the void type, used as the result of operators that branch away, or do not return anything.

fn bool_or_int(&mut self, span: Span) -> Self::TypeVariable

Create a type variable that may be either a boolean or an integer.

fn any_t(&mut self, span: Span) -> Self::TypeVariable

Create a type variable that can be any type T.

Each use of any_t by the same operator refers to the same type variable.

Loading content...