Module peepmatic_runtime::integer_interner

Interner for (potentially large) integer values.

We support matching on integers that can be represented by u64, but only support automata results that fit in a u32. So we intern the (relatively few compared to the full range of u64) integers we are matching against here and then reference them by IntegerId.

Structs

IntegerId
An identifier for an interned integer.

IntegerInterner
An interner for integer values.